♻️ 時刻表示部分をコンポーネント化
