+++ src/App.jsx
@@ -1,21 +1,6 @@
 import { useInterval } from "use-interval";
 import { useCallback, useEffect, useState } from "react";
-
-const getMinutesStr = (time) => {
-  return Math.floor(time / 60)
-    .toString()
-    .padStart(2, "0");
-};
-
-const getSecondsStr = (time) => {
-  return Math.floor(time % 60)
-    .toString()
-    .padStart(2, "0");
-};
-
-const getCentiSecondsStr = (time) => {
-  return time.toFixed(2).split(".")[1];
-};
+import TimeDisplay from "src/components/TimeDisplay";
 
 const App = () => {
   const [running, setRunning] = useState(false);
@@ -51,13 +36,7 @@ const App = () => {
 
   return (
     <div className="antialiased h-screen flex flex-col justify-center items-center dark:bg-gray-900">
-      <div className="my-6 flex font-roboto text-9xl text-gray-800 dark:text-gray-100">
-        <div className="mx-3">{getMinutesStr(time)}</div>
-        <div className="-translate-y-2.5">:</div>
-        <div className="mx-3">{getSecondsStr(time)}</div>
-        <div>.</div>
-        <div className="mx-3">{getCentiSecondsStr(time)}</div>
-      </div>
+      <TimeDisplay time={time} />
       <div className="my-6 flex items-center">
         <button
           className="w-24 h-24 mx-8 text-base font-bold text-gray-600 dark:text-gray-200 border-2 border-gray-400 dark:border-gray-600 bg-gray-100 dark:bg-gray-700 rounded-full hover:bg-gray-200 dark:hover:bg-gray-600 active:translate-y-0.5"
